Created parametrize test for DomriaOutputConvertor
