+++ docker-compose.yaml
@@ -80,7 +80,7 @@ services:
     image: redis:6.2
     ports:
       - "6379:6379"
- 
+
   chromedriver:
     profiles: ["dev", "dev-python"]
     image: selenium/standalone-chrome
