+++ tests/convertors.py
@@ -4,7 +4,8 @@ Model convertors
 from abc import ABC, abstractmethod
 from typing import List, Dict
 
-from service_api.models import State, RealtyType, OperationType, City, Category, Realty, RealtyDetails, Service
+from service_api.models import State, RealtyType, OperationType, City, Category, Realty, RealtyDetails, Service, \
+    StateAlias, RealtyTypeAlias, OperationTypeAlias, CityAlias
 
 
 class AbstractModelConvertor(ABC):
@@ -88,6 +89,55 @@ class CityModelConvertor(AbstractModelConvertor):
         """
         cities_data, *_ = args
         return [City(**city) for city in cities_data]
+
+
+class StateAliasModelConvertor(AbstractModelConvertor):
+    """
+    StateAlias Model Convertor
+    """
+    def convert_to_model(self, *args):
+        """
+        Converts container of deserialized objects to the StateAlias models
+        """
+        state_alias_data, *_ = args
+        return [StateAlias(**state_alias) for state_alias in state_alias_data]
+
+
+class RealtyTypeAliasModelConvertor(AbstractModelConvertor):
+    """
+    StateAlias Model Convertor
+    """
+    def convert_to_model(self, *args):
+        """
+        Converts container of deserialized objects to the RealtyTypeAlias models
+        """
+        realty_type_alias_data, *_ = args
+        return [RealtyTypeAlias(**realty_type_alias) for realty_type_alias in realty_type_alias_data]
+
+
+class OperationTypeAliasModelConvertor(AbstractModelConvertor):
+    """
+    StateAlias Model Convertor
+    """
+    def convert_to_model(self, *args):
+        """
+        Converts container of deserialized objects to the OperationTypeAlias models
+        """
+        operation_type_alias_data, *_ = args
+        return [OperationTypeAlias(**operation_type_alias) for operation_type_alias in operation_type_alias_data]
+
+
+class CityAliasModelConvertor(AbstractModelConvertor):
+    """
+    StateAlias Model Convertor
+    """
+    def convert_to_model(self, *args):
+        """
+        Converts container of deserialized objects to the CityAlias models
+        """
+        city_alias_data, *_ = args
+        return [CityAlias(**city_alias) for city_alias in city_alias_data]
+
 
 
 class RealtyDetailsModelConvertor(AbstractModelConvertor):
